requirements.txt
pyalsaaudio==0.8.4
paho-mqtt==1.3.1
ruamel.yaml==0.16.10